feature: refactor to compile (#3)
