+++ command_api/src/main/java/com/meeting/sport/app/sport_field/command/AddSportFieldCommand.java
@@ -1,12 +1,10 @@
 package com.meeting.sport.app.sport_field.command;
 
 import com.meeting.sport.app.Command;
-import com.meeting.sport.app.sport_field.FieldSpace;
-import com.meeting.sport.app.sport_field.FieldType;
 
 public record AddSportFieldCommand(
-        FieldType fieldType,
-        FieldSpace fieldSpace,
+        String fieldType,
+        String fieldSpace,
         String city,
         String street,
         String number
