+++ infrastructure/src/main/java/com/meeting/sport/app/dto/EventRoleResponse.java
@@ -1,13 +1,12 @@
 package com.meeting.sport.app.dto;
 
-import com.meeting.sport.app.sport_event.GameRole;
-import com.meeting.sport.app.sport_event.SportEvent;
 import lombok.Builder;
 
 @Builder
 public record EventRoleResponse(
         Long id,
-        GameRole gameRole,
+
+        String  gameRole,
         boolean isAvailable,
         Long userId,
