+++ infrastructure/src/main/java/com/meeting/sport/app/dto/SportFieldResponse.java
@@ -1,14 +1,12 @@
 package com.meeting.sport.app.dto;
 
-import com.meeting.sport.app.sport_field.FieldSpace;
-import com.meeting.sport.app.sport_field.FieldType;
 import lombok.Builder;
 
 @Builder
 public record SportFieldResponse(
         Long id,
-        FieldType fieldType,
-        FieldSpace fieldSpace,
+        String fieldType,
+        String fieldSpace,
         String city,
         String street,
         String number
