+++ infrastructure/src/main/java/com/meeting/sport/app/sport_event/EventRoleRepositoryImpl.java
@@ -2,7 +2,6 @@ package com.meeting.sport.app.sport_event;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @Repository
 class EventRoleRepositoryImpl implements EventRoleRepository {
@@ -16,19 +15,19 @@ class EventRoleRepositoryImpl implements EventRoleRepository {
 
     @Override
     public Long save(EventRole eventRole) {
-        EventRoleEntity entity = eventRoleRepositoryJPA.save(EventRoleMapper1.modelToEntity(eventRole));
+        EventRoleEntity entity = eventRoleRepositoryJPA.save(EventRoleMapper.modelToEntity(eventRole));
         return entity.getId();
     }
 
     @Override
     public EventRole getEventRoleByUserAndEvent(Long userId,Long eventId) {
         EventRoleEntity eventRole = eventRoleRepositoryJPA.getEventRoleEntityByUserIdAndSportEventEntityId(userId,eventId);
-        return EventRoleMapper1.entityToModel(eventRole);
+        return EventRoleMapper.entityToModel(eventRole);
     }
 
     @Override
     public List<EventRole> getEventRoleEntitiesByUserEntityId(Long userId) {
         List<EventRoleEntity> eventRoleEntities = eventRoleRepositoryJPA.getEventRoleEntitiesByUserId(userId);
-        return eventRoleEntities.stream().map(EventRoleMapper1::entityToModel).toList();
+        return eventRoleEntities.stream().map(EventRoleMapper::entityToModel).toList();
     }
 }
