+++ infrastructure/src/main/java/com/meeting/sport/app/token/TokenMapper.java
@@ -1,18 +1,57 @@
 package com.meeting.sport.app.token;
 
 import com.meeting.sport.app.user.Token;
+import com.meeting.sport.app.user.User;
 import com.meeting.sport.app.user.UserMapper;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-@Mapper(componentModel = "spring",uses = {UserMapper.class})
-public interface TokenMapper {
+import com.meeting.sport.app.user.dto.UserDTO;
 
-    @Mapping(source = "userEntity", target = "userDTO")
-    TokenDTO entityToDTO(TokenEntity entity);
+public class TokenMapper {
 
-    @Mapping(source = "userDTO", target = "user")
-    Token DTOToModel(TokenDTO dto);
+    public static TokenDTO entityToDTO(TokenEntity entity) {
+        if (entity == null) {
+            return null;
+        }
 
-    @Mapping(source = "user", target = "userEntity")
-    TokenEntity modelToEntity(Token token);
+        UserDTO userDTO = UserMapper.entityToDTO(entity.getUserEntity());
+        Long id = entity.getId();
+        String token = entity.getToken();
+        boolean revoked = entity.isRevoked();
+        boolean expired = entity.isExpired();
+
+        return new TokenDTO(id, token, revoked, expired, userDTO);
+
+    }
+
+    public static Token DTOToModel(TokenDTO dto) {
+        if (dto == null) {
+            return null;
+        }
+
+
+        User user = UserMapper.DTOToModel(dto.userDTO());
+        Long id = dto.id();
+        String token = dto.token();
+        boolean revoked = dto.revoked();
+        boolean expired = dto.expired();
+
+        return new Token(id, token, revoked, expired, user);
+
+    }
+
+
+    public static TokenEntity modelToEntity(Token token) {
+        if (token == null) {
+            return null;
+        }
+
+        TokenEntity.TokenEntityBuilder tokenEntity = TokenEntity.builder();
+
+        tokenEntity.userEntity(UserMapper.modelToEntity(token.getUser()));
+        tokenEntity.id(token.getId());
+        tokenEntity.token(token.getToken());
+        tokenEntity.revoked(token.isRevoked());
+        tokenEntity.expired(token.isExpired());
+
+        return tokenEntity.build();
+    }
 }
