--- infrastructure/src/main/java/com/meeting/sport/app/user/AuthorityMapper1.java
@@ -1,4 +1,0 @@
-package com.meeting.sport.app.user;
-
-public class AuthorityMapper1 {
-}
