+++ infrastructure/src/main/resources/changelog-master.xml
@@ -1,0 +1,10 @@
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+        http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.4.xsd">
+
+        <include file="/db-changelog.xml"/>
+        <include file="/insert-changelog.xml"/>
+
+</databaseChangeLog>
