+++ service/src/main/java/com/meeting/sport/app/user/QueryUserServiceImpl.java
@@ -1,16 +1,12 @@
 package com.meeting.sport.app.user;
 
-import com.meeting.sport.app.sport_event.SportEventFacade;
-import com.meeting.sport.app.user.dto.UserDTO;
 import com.meeting.sport.app.user.dto.UserResponse;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import java.util.Optional;
-
 @Component
 @AllArgsConstructor
-public class QueryUserServiceImpl implements QueryUserService {
+class QueryUserServiceImpl implements QueryUserService {
 
     private final UserRepository userRepository;
     private final UserService userService;
@@ -23,6 +19,6 @@ public class QueryUserServiceImpl implements QueryUserService {
 //    }
     public UserResponse getUserResponse(Long id){
         User user = userService.getLoggedUser(id);
-        return UserMapper1.modelToResponse(user);
+        return UserMapper.modelToResponse(user);
     }
 }
