+++ service/src/main/java/com/meeting/sport/app/user/UserServiceImpl.java
@@ -8,21 +8,20 @@ import org.springframework.stereotype.Component;
 
 @Component
 @AllArgsConstructor
-class UserServiceImpl implements UserService{
+class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
-    private final UserMapper userMapper;
     private final PasswordEncoder passwordEncoder;
 
     @Override
-    public User getLoggedUser(String email){
+    public User getLoggedUser(String email) {
 
-        return userMapper.DTOToModel(userRepository.findUserByEmail(email));
+        return UserMapper.DTOToModel(userRepository.findUserByEmail(email));
     }
 
     @Override
     public User getLoggedUser(Long id) {
-        return userMapper.DTOToModel(userRepository.findById(id));
+        return UserMapper.DTOToModel(userRepository.findById(id));
     }
 
     @Override
@@ -30,15 +29,18 @@ class UserServiceImpl implements UserService{
         return userRepository.findUserByEmail(email);
     }
 
-    public UserDTO createUser(RegisterRequest request){
-        UserDTO user = UserDTO.builder()
-                .firstname(request.firstName())
-                .lastname(request.lastName())
-                .age(request.age())
-                .email(request.email())
-                .password(passwordEncoder.encode(request.password()))
-                .role(Role.USER)
-                .build();
+    public UserDTO createUser(RegisterRequest request) {
+
+        String encodedPassword = passwordEncoder.encode(request.password());
+
+        User user = User.createUser(
+                request.firstName(),
+                request.lastName(),
+                request.age(),
+                request.email(),
+                encodedPassword
+        );
+
         return userRepository.saveUser(user);
     }
 }
