+++ webapp/webapp/pages/table-row/table-row.component.html
@@ -1,6 +1,6 @@
 <section>
   <div class="example-button-row">
-    <button mat-raised-button>Profile</button>
+    <button mat-raised-button routerLink="/profile" routerLinkActive="active">Profile</button>
     <button mat-raised-button routerLink="/events" routerLinkActive="active">Events</button>
     <button mat-raised-button routerLink="/add-event" routerLinkActive="active">Add Event</button>
   </div>
