+++ webapp/webapp/src/app/app.routes.ts
@@ -8,6 +8,7 @@ import {AddEventComponent} from "../../pages/add-event/add-event.component";
 import {SelectedEventCardComponent} from "../../pages/selected-event-card/selected-event-card.component";
 import {AddEventRoleComponent} from "../../pages/add-event-role/add-event-role.component";
 import {AddFieldComponent} from "../../pages/add-field/add-field.component";
+import {UserProfileComponent} from "../../pages/user-profile/user-profile.component";
 
 export const routes: Routes = [
   {path: "login", component: LoginComponent},
@@ -17,7 +18,8 @@ export const routes: Routes = [
   {path: "add-event", component: AddEventComponent},
   {path: "event/:eventId", component: SelectedEventCardComponent},
   {path: "event/:eventId/add-event-role", component: AddEventRoleComponent},
-  {path: "event/:eventId/add-field", component: AddFieldComponent}
+  {path: "event/:eventId/add-field", component: AddFieldComponent},
+  {path: "profile", component: UserProfileComponent}
 ];
 @NgModule({
   imports:[RouterModule.forRoot(routes)],
